Refactor foreground style customization APIs
