+++ Sources/MarkdownView/Configurations/ForegroundStyle/AnyForegroundStyleGroup.swift
@@ -1,7 +1,7 @@
 import SwiftUI
 
-/// A type-erased MarkdownForegroundStyleGroup value.
-public struct AnyMarkdownForegroundStyleGroup: Sendable {
+/// A type-erased `HeadingStyleGroup` value.
+public struct AnyHeadingStyleGroup: Sendable {
     var _h1: AnyShapeStyle
     var _h2: AnyShapeStyle
     var _h3: AnyShapeStyle
@@ -9,7 +9,7 @@ public struct AnyMarkdownForegroundStyleGroup: Sendable {
     var _h5: AnyShapeStyle
     var _h6: AnyShapeStyle
     
-    init(_ group: some MarkdownForegroundStyleGroup) {
+    init(_ group: some HeadingStyleGroup) {
         _h1 = AnyShapeStyle(group.h1)
         _h2 = AnyShapeStyle(group.h2)
         _h3 = AnyShapeStyle(group.h3)
@@ -19,7 +19,7 @@ public struct AnyMarkdownForegroundStyleGroup: Sendable {
     }
 }
 
-extension AnyMarkdownForegroundStyleGroup: MarkdownForegroundStyleGroup {
+extension AnyHeadingStyleGroup: HeadingStyleGroup {
     public var h1: AnyShapeStyle { _h1 }
     public var h2: AnyShapeStyle { _h2 }
     public var h3: AnyShapeStyle { _h3 }
@@ -28,4 +28,4 @@ extension AnyMarkdownForegroundStyleGroup: MarkdownForegroundStyleGroup {
     public var h6: AnyShapeStyle { _h6 }
 }
 
-extension AnyMarkdownForegroundStyleGroup: Equatable { }
+extension AnyHeadingStyleGroup: Equatable { }
