+++ Sources/MarkdownView/Configurations/ForegroundStyle/AutomaticForegroundStyleGroup.swift
@@ -2,11 +2,10 @@ import SwiftUI
 
 /// A foreground style group that resolves its content appearance automatically based on the current context.
 ///
-/// Use ``MarkdownView/MarkdownForegroundStyleGroup/automatic`` to construct this type.
-public struct AutomaticForegroundStyleGroup: MarkdownForegroundStyleGroup { }
+/// Use ``HeadingForegroundStyleGroup/automatic`` to construct this type.
+public struct AutomaticHeadingForegroundStyleGroup: HeadingStyleGroup { }
 
-extension MarkdownForegroundStyleGroup where Self == AutomaticForegroundStyleGroup {
+extension HeadingStyleGroup where Self == AutomaticHeadingForegroundStyleGroup {
     /// A foreground style group that resolves its content appearance automatically based on the current context.
-    static public var automatic: AutomaticForegroundStyleGroup { .init() }
+    static public var automatic: AutomaticHeadingForegroundStyleGroup { .init() }
 }
-
