+++ Sources/MarkdownView/Documentation.docc/Detailed Documentations/HeadingForegroundStyleGroup.md
@@ -1,0 +1,24 @@
+# ``HeadingForegroundStyleGroup``
+
+## Topics
+
+### Getting built-in style group
+
+- ``HeadingForegroundStyleGroup/automatic``
+
+### Creating foreground style groups
+
+- ``HeadingForegroundStyleGroup/h1``
+- ``HeadingForegroundStyleGroup/h2``
+- ``HeadingForegroundStyleGroup/h3``
+- ``HeadingForegroundStyleGroup/h4``
+- ``HeadingForegroundStyleGroup/h5``
+- ``HeadingForegroundStyleGroup/h6``
+
+### Supporting types
+
+- ``AutomaticHeadingForegroundStyleGroup``
+
+### Substituting style group
+
+- ``AnyHeadingForegroundStyleGroup``
