--- Sources/MarkdownView/Documentation.docc/Detailed Documentations/MarkdownForegroundStyleGroup.md
@@ -1,28 +1,0 @@
-# ``MarkdownForegroundStyleGroup``
-
-## Topics
-
-### Getting built-in style group
-
-- ``MarkdownForegroundStyleGroup/automatic``
-
-### Creating foreground style groups
-
-- ``MarkdownForegroundStyleGroup/h1-8qmk8``
-- ``MarkdownForegroundStyleGroup/h2-4qcaq``
-- ``MarkdownForegroundStyleGroup/h3-3bblc``
-- ``MarkdownForegroundStyleGroup/h4-7liif``
-- ``MarkdownForegroundStyleGroup/h5-9y0ap``
-- ``MarkdownForegroundStyleGroup/h6-53rwk``
-- ``MarkdownForegroundStyleGroup/blockQuote-9ika``
-- ``MarkdownForegroundStyleGroup/codeBlock-71cpp``
-- ``MarkdownForegroundStyleGroup/tableHeader-6w40j``
-- ``MarkdownForegroundStyleGroup/tableBody-8q0gs``
-
-### Supporting types
-
-- ``AutomaticForegroundStyleGroup``
-
-### Substituting style group
-
-- ``AnyMarkdownForegroundStyleGroup``
